Restructured all my docker images
So basically applant/mruby-cli uses a version of Ubuntu that is no
longer supported, so I took this as a hint that it's time to rebuild my
entire docker infrastructure.

This now means that all export targets are individual docker images,
which will bulk up the download requirement for someone exporting all
versions but will make it smaller for someone just doing one or two
export targets.