Release v0.3.12.1
Web exports have had issues since v0.3.8 with regards to speed but in
the last release has also gained a terrible RAM hog issue where it
spikes to 8GB before settling back down. I've managed to fix the RAM
spike but the speed issue still kind of exists, it's gone from a 80%
drop in speed to just a 20% drop, which I don't love but I'm not sure
what's going on so for now, it's a fix at least.