Rework the compilation instructions
Changed some wordings, and changed the instruction for adding
Taylor to PATH too, since `~/.local/bin` isn't guaranteed to be present in every
system, exporting the path automatically on the `.bashrc` is far more "universal" & seems
like the convention for adding programs to PATH.

Other little change is that the Docker section is now outside the compile section. The reason
is that I think it is notable enough for it to be separated from the other OS-specific instructions.