Fix specs in browser to work a bit more sensibly
Basically I was trying to get audio to play automatically in automated
testing but failed to get it to work. In the process I noticed that the
tests didn't follow best practices for web (and still kind of don't
since it's all run in one iteration of the main loop) but it does do
more sensible things by actually using the main loop.